1) Changed all log entries to be of the form logger.info("The new entry is {}.", entry); Before the change, they were of the form logger.info("The new entry is "+entry+".");  The first form executes faster; here is excerpt from http://logback.qos.ch/manual/architecture.html “The following two lines will yield the exact same output. However, in case of a disabled logging statement, the second variant will outperform the first variant by a factor of at least 30.
logger.debug("The new entry is "+entry+".");
logger.debug("The new entry is {}.", entry);”

2) added to the log IP address of the client (it is extracted from
HttpServletRequest);

3) added requestsUniqueId field to NdexService class. This field is
obtained by  concatenation of (Timestamp + Thread Id);
The format of this field is "yyyy-MM-dd-HH-mm-ss-SSS-thread_id", 
for example:  2015-06-25-12-21-436-30.  This unique Id is the Id that
will never be reused in future, nor was it used in the past.
It will be used to stamp all requests in the log file.